# swagger-spec.json
{"openapi":"3.0.0","info":{"title":"DBLab","description":"DB Lab #2","version":"1.0","contact":{}},"tags":[{"name":"database","description":""}],"servers":[],"components":{"securitySchemes":{"JWT":{"scheme":"bearer","bearerFormat":"JWT","type":"http","in":"header"}},"schemas":{"CreateBookDto":{"type":"object","properties":{}},"CreateUserDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"},"phoneNumber":{"type":"string"},"books":{"type":"array","items":{"type":"string"}}},"required":["username","password","email","role","phoneNumber","books"]},"CreateGenreDto":{"type":"object","properties":{}},"CreateBlogPostDto":{"type":"object","properties":{"title":{"type":"string"},"date":{"format":"date-time","type":"string"},"text":{"type":"string"}},"required":["title","date","text"]},"EditBlogPostDto":{"type":"object","properties":{"title":{"type":"string"},"text":{"type":"string"}},"required":["title","text"]},"CreateProjectDto":{"type":"object","properties":{"title":{"type":"string"},"deadline":{"format":"date-time","type":"string"},"type":{"type":"string"},"size":{"type":"string"},"description":{"type":"string"},"skillGuarantee":{"type":"boolean"},"subject":{"type":"string"}},"required":["title","deadline","type","size","description","skillGuarantee","subject"]},"CreateResumeDto":{"type":"object","properties":{"name":{"type":"string"},"text":{"type":"string"}},"required":["name","text"]},"CreateRateDto":{"type":"object","properties":{"rate":{"type":"number"},"ratedID":{"type":"number"}},"required":["rate","ratedID"]}}},"paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/book/post":{"post":{"operationId":"BooksController_postGenre","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookDto"}}}},"responses":{"401":{"description":"Unauthorized"}},"security":[{"JWT":[]}]}},"/book":{"get":{"operationId":"BooksController_getAll","parameters":[],"responses":{"200":{"description":""}}}},"/book/delete/{id}":{"delete":{"operationId":"BooksController_deleteBook","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"401":{"description":"Unauthorized"}},"security":[{"JWT":[]}]}},"/book/put/{id}":{"put":{"operationId":"BooksController_editBook","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookDto"}}}},"responses":{"401":{"description":"Unauthorized"}},"security":[{"JWT":[]}]}},"/users/post":{"post":{"operationId":"UserController_postUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}}}},"/users":{"get":{"operationId":"UserController_getAll","parameters":[],"responses":{"401":{"description":"Unauthorized"}},"security":[{"JWT":[]}]}},"/users/books":{"get":{"operationId":"UserController_getBooks","parameters":[],"responses":{"401":{"description":"Unauthorized"}},"security":[{"JWT":[]}]}},"/users/delete/{id}":{"delete":{"operationId":"UserController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"401":{"description":"Unauthorized"}},"security":[{"JWT":[]}]}},"/users/put/{id}":{"put":{"operationId":"UserController_edituser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"401":{"description":"Unauthorized"}},"security":[{"JWT":[]}]}},"/genre/post":{"post":{"operationId":"GenreController_postGenre","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGenreDto"}}}},"responses":{"401":{"description":"Unauthorized"}},"security":[{"JWT":[]}]}},"/genre":{"get":{"operationId":"GenreController_getAll","parameters":[],"responses":{"200":{"description":""}}}},"/genre/delete/{id}":{"delete":{"operationId":"GenreController_deleteGenre","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"401":{"description":"Unauthorized"}},"security":[{"JWT":[]}]}},"/genre/put/{id}":{"put":{"operationId":"GenreController_edituser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGenreDto"}}}},"responses":{"401":{"description":"Unauthorized"}},"security":[{"JWT":[]}]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"responses":{"201":{"description":""}}}},"/auth/profile":{"get":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":""}},"security":[{"JWT":[]}]}},"/jobseekers/blogPost/{id}":{"get":{"operationId":"JobseekersController_getBlogPost","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource"},"404":{"description":"Resource was not found"}},"security":[{"JWT":[]}]},"put":{"operationId":"JobseekersController_editBlogPost","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditBlogPostDto"}}}},"responses":{"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource"},"404":{"description":"Resource was not found"}},"security":[{"JWT":[]}]},"delete":{"operationId":"JobseekersController_deleteBlogPost","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource"},"404":{"description":"Resource was not found"}},"security":[{"JWT":[]}]}},"/jobseekers/blogPost":{"get":{"operationId":"JobseekersController_getAll","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"JobseekersController_addBlogPost","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBlogPostDto"}}}},"responses":{"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource"}},"security":[{"JWT":[]}]}},"/jobseekers/project/{id}":{"get":{"operationId":"JobseekersController_getProject","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource"},"404":{"description":"Resource was not found"}},"security":[{"JWT":[]}]},"put":{"operationId":"JobseekersController_editProject","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectDto"}}}},"responses":{"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource"},"404":{"description":"Resource was not found"}},"security":[{"JWT":[]}]},"delete":{"operationId":"JobseekersController_deleteProject","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource"},"404":{"description":"Resource was not found"}},"security":[{"JWT":[]}]}},"/jobseekers/project":{"post":{"operationId":"JobseekersController_addProject","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectDto"}}}},"responses":{"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource"}},"security":[{"JWT":[]}]}},"/jobseekers/resume":{"get":{"operationId":"JobseekersController_getResume","parameters":[],"responses":{"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource"},"404":{"description":"Resource was not found"}},"security":[{"JWT":[]}]},"post":{"operationId":"JobseekersController_addResume","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateResumeDto"}}}},"responses":{"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource"}},"security":[{"JWT":[]}]},"put":{"operationId":"JobseekersController_editResume","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateResumeDto"}}}},"responses":{"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource"},"404":{"description":"Resource was not found"}},"security":[{"JWT":[]}]},"delete":{"operationId":"JobseekersController_deleteResume","parameters":[],"responses":{"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource"},"404":{"description":"Resource was not found"}},"security":[{"JWT":[]}]}},"/jobseekers/rate/{ratedID}":{"get":{"operationId":"JobseekersController_getRate","parameters":[{"name":"ratedID","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource"},"404":{"description":"Resource was not found"}},"security":[{"JWT":[]}]},"delete":{"operationId":"JobseekersController_deleteRate","parameters":[{"name":"ratedID","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource"},"404":{"description":"Resource was not found"}},"security":[{"JWT":[]}]}},"/jobseekers/rate":{"post":{"operationId":"JobseekersController_addRate","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRateDto"}}}},"responses":{"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource"}},"security":[{"JWT":[]}]},"put":{"operationId":"JobseekersController_editRate","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRateDto"}}}},"responses":{"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource"},"404":{"description":"Resource was not found"}},"security":[{"JWT":[]}]}}}}